Refactor case working copy state into its own file
