+++ plugin-hrm-form/src/___tests__/states/case/reducer.test.ts
@@ -1,15 +1,20 @@
 import { DefinitionVersionId, loadDefinition } from 'hrm-form-definitions';
 
-import { CaseState, CaseWorkingCopy, reduce } from '../../../states/case/reducer';
+import { reduce } from '../../../states/case/reducer';
 import * as actions from '../../../states/case/actions';
 import * as GeneralActions from '../../../states/actions';
 import { Case, CaseItemEntry } from '../../../types/types';
 import { REMOVE_CONTACT_STATE } from '../../../states/types';
-import { TemporaryCaseInfo, UPDATE_CASE_CONTACT } from '../../../states/case/types';
+import { CaseState, CaseWorkingCopy, TemporaryCaseInfo, UPDATE_CASE_CONTACT } from '../../../states/case/types';
 import { CaseItemAction } from '../../../states/routing/types';
 import { connectedCaseBase, namespace, RootState } from '../../../states';
 import { householdSectionApi } from '../../../states/case/sections/household';
 import { CaseSectionApi } from '../../../states/case/sections/api';
+import {
+  initialiseCaseSectionWorkingCopy,
+  removeCaseSectionWorkingCopy,
+  updateCaseSectionWorkingCopy,
+} from '../../../states/case/caseWorkingCopy';
 
 const task = { taskSid: 'task1' };
 
@@ -268,277 +273,6 @@ describe('test reducer', () => {
           },
         }),
       ).toThrow();
-    });
-  });
-  describe('Working copy reducers', () => {
-    const baselineDate = new Date(2022, 1, 28);
-    const updatedSection: CaseItemEntry = {
-      id: 'existingHousehold',
-      twilioWorkerId: 'other-worker-sid',
-      form: { otherProp: 'other-value' },
-      createdAt: baselineDate.toISOString(),
-    };
-
-    let stubApi: CaseSectionApi<any>;
-    const stubUpdateWorkingCopy: CaseWorkingCopy = { sections: { mock: { existing: { mockId: updatedSection } } } };
-
-    beforeEach(() => {
-      stubApi = {
-        ...householdSectionApi,
-        updateWorkingCopy: jest.fn().mockReturnValue(stubUpdateWorkingCopy),
-      };
-    });
-
-    describe('UPDATE_CASE_WORKING_COPY', () => {
-      test('Specifies id - updates case working copy using APIs updateWorkingCopy function', () => {
-        const initialState: CaseState = {
-          tasks: {
-            task1: {
-              ...state.tasks.task1,
-              caseWorkingCopy: {
-                sections: {
-                  households: {
-                    existing: {
-                      existingHousehold: {
-                        id: 'existingHousehold',
-                        twilioWorkerId: 'worker-sid',
-                        form: { prop: 'value' },
-                        createdAt: baselineDate.toISOString(),
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        };
-
-        const expected: CaseState = {
-          tasks: {
-            task1: {
-              ...state.tasks.task1,
-              caseWorkingCopy: stubUpdateWorkingCopy,
-            },
-          },
-        };
-
-        const result = reduce(
-          initialState,
-          actions.updateCaseSectionWorkingCopy(task.taskSid, stubApi, updatedSection, 'existingHousehold'),
-        );
-        expect(result).toStrictEqual(expected);
-        expect(stubApi.updateWorkingCopy).toHaveBeenCalledWith(
-          initialState.tasks.task1.caseWorkingCopy,
-          updatedSection,
-          'existingHousehold',
-        );
-      });
-      test('Specifies no id - updates case working copy using APIs updateWorkingCopy function without setting an id', () => {
-        const initialState: CaseState = {
-          tasks: {
-            task1: {
-              ...state.tasks.task1,
-              caseWorkingCopy: {
-                sections: {},
-              },
-            },
-          },
-        };
-
-        const expected: CaseState = {
-          tasks: {
-            task1: {
-              ...state.tasks.task1,
-              caseWorkingCopy: stubUpdateWorkingCopy,
-            },
-          },
-        };
-
-        const result = reduce(
-          initialState,
-          actions.updateCaseSectionWorkingCopy(task.taskSid, stubApi, updatedSection),
-        );
-        expect(result).toStrictEqual(expected);
-        expect(stubApi.updateWorkingCopy).toHaveBeenCalledWith(
-          initialState.tasks.task1.caseWorkingCopy,
-          updatedSection,
-          undefined,
-        );
-      });
-    });
-
-    describe('INIT_CASE_WORKING_COPY', () => {
-      test('Specifies id that exists in case - updates case working copy using APIs updateWorkingCopy function, copying the section from the connected case', () => {
-        const initialState: CaseState = {
-          tasks: {
-            task1: {
-              ...state.tasks.task1,
-              connectedCase: {
-                ...state.tasks.task1.connectedCase,
-                info: {
-                  ...state.tasks.task1.connectedCase.info,
-                  households: [
-                    {
-                      id: 'existingHousehold',
-                      twilioWorkerId: 'other-worker-sid',
-                      household: { otherProp: 'other-value' },
-                      createdAt: baselineDate.toISOString(),
-                    },
-                  ],
-                },
-              },
-              caseWorkingCopy: {
-                sections: {},
-              },
-            },
-          },
-        };
-
-        const expected: CaseState = {
-          tasks: {
-            task1: {
-              ...initialState.tasks.task1,
-              caseWorkingCopy: stubUpdateWorkingCopy,
-            },
-          },
-        };
-
-        const result = reduce(
-          initialState,
-          actions.initialiseCaseSectionWorkingCopy(task.taskSid, stubApi, 'existingHousehold'),
-        );
-        expect(result).toStrictEqual(expected);
-        expect(stubApi.updateWorkingCopy).toHaveBeenCalledWith(
-          initialState.tasks.task1.caseWorkingCopy,
-          updatedSection,
-          'existingHousehold',
-        );
-      });
-
-      test('Specifies no id - updates case working copy using APIs updateWorkingCopy function specifying a blank CaseItemEntry with a random ID', () => {
-        const initialState: CaseState = {
-          tasks: {
-            task1: {
-              ...state.tasks.task1,
-              caseWorkingCopy: {
-                sections: {},
-              },
-            },
-          },
-        };
-
-        const expected: CaseState = {
-          tasks: {
-            task1: {
-              ...initialState.tasks.task1,
-              caseWorkingCopy: stubUpdateWorkingCopy,
-            },
-          },
-        };
-
-        const result = reduce(initialState, actions.initialiseCaseSectionWorkingCopy(task.taskSid, stubApi));
-        expect(result).toStrictEqual(expected);
-        expect(stubApi.updateWorkingCopy).toHaveBeenCalledWith(
-          initialState.tasks.task1.caseWorkingCopy,
-          { id: expect.any(String), form: {}, createdAt: null, twilioWorkerId: null },
-          undefined,
-        );
-      });
-
-      test('Specifies id when section not present in the connected case - throws', () => {
-        const initialState: CaseState = {
-          tasks: {
-            task1: {
-              ...state.tasks.task1,
-              caseWorkingCopy: {
-                sections: {},
-              },
-            },
-          },
-        };
-
-        expect(() =>
-          reduce(initialState, actions.initialiseCaseSectionWorkingCopy(task.taskSid, stubApi, 'existingHousehold')),
-        ).toThrow();
-      });
-
-      test('Specifies id that is not present in the connected case - throws', () => {
-        const initialState: CaseState = {
-          tasks: {
-            task1: {
-              ...state.tasks.task1,
-              connectedCase: {
-                ...state.tasks.task1.connectedCase,
-                info: {
-                  ...state.tasks.task1.connectedCase.info,
-                  households: [],
-                },
-              },
-              caseWorkingCopy: {
-                sections: {},
-              },
-            },
-          },
-        };
-
-        expect(() =>
-          reduce(initialState, actions.initialiseCaseSectionWorkingCopy(task.taskSid, stubApi, 'existingHousehold')),
-        ).toThrow();
-      });
-    });
-
-    describe('REMOVE_CASE_SECTION_WORKING_COPY', () => {
-      test("Task doesn't exist - noop", () => {
-        const initialState: CaseState = {
-          tasks: {},
-        };
-
-        const resultWithNoId = reduce(initialState, actions.removeCaseSectionWorkingCopy('non existent task', stubApi));
-        expect(resultWithNoId).toStrictEqual(initialState);
-
-        const resultWithId = reduce(
-          initialState,
-          actions.removeCaseSectionWorkingCopy('non existent task', stubApi, 'non existent id'),
-        );
-        expect(resultWithId).toStrictEqual(initialState);
-      });
-      test("Task exists - calls api's updateWorkingCopy with id and an undefined item", () => {
-        const initialState: CaseState = {
-          tasks: {
-            task1: {
-              ...state.tasks.task1,
-              caseWorkingCopy: {
-                sections: {},
-              },
-            },
-          },
-        };
-
-        const expected: CaseState = {
-          tasks: {
-            task1: {
-              ...initialState.tasks.task1,
-              caseWorkingCopy: stubUpdateWorkingCopy,
-            },
-          },
-        };
-
-        const resultWithoutId = reduce(initialState, actions.removeCaseSectionWorkingCopy(task.taskSid, stubApi));
-        expect(resultWithoutId).toStrictEqual(expected);
-        expect(stubApi.updateWorkingCopy).toHaveBeenCalledWith(
-          initialState.tasks.task1.caseWorkingCopy,
-          undefined,
-          undefined,
-        );
-
-        const resultWithId = reduce(initialState, actions.removeCaseSectionWorkingCopy(task.taskSid, stubApi, 'an id'));
-        expect(resultWithId).toStrictEqual(expected);
-        expect(stubApi.updateWorkingCopy).toHaveBeenCalledWith(
-          initialState.tasks.task1.caseWorkingCopy,
-          undefined,
-          'an id',
-        );
-      });
     });
   });
 });
