+++ plugin-hrm-form/src/components/case/Case.tsx
@@ -91,7 +91,6 @@ const Case: React.FC<Props> = ({
   removeConnectedCase,
   updateCaseInfo,
   updateCaseStatus,
-  markCaseAsUpdated,
   changeRoute,
   isCreating,
   handleClose,
@@ -464,7 +463,6 @@ const mapDispatchToProps = dispatch => {
     updateTempInfo: bindActionCreators(CaseActions.updateTempInfo, dispatch),
     updateCaseStatus: bindActionCreators(CaseActions.updateCaseStatus, dispatch),
     setConnectedCase: bindActionCreators(CaseActions.setConnectedCase, dispatch),
-    markCaseAsUpdated: bindActionCreators(CaseActions.markCaseAsUpdated, dispatch),
     updateDefinitionVersion: bindActionCreators(ConfigActions.updateDefinitionVersion, dispatch),
     releaseContacts: bindActionCreators(ContactActions.releaseContacts, dispatch),
     loadRawContacts: bindActionCreators(ContactActions.loadRawContacts, dispatch),
